Refactor by extracting modal to own component
