+++ packages/customer-effort-score/package.json
@@ -20,7 +20,11 @@
 	"module": "build-module/index.js",
 	"react-native": "src/index",
 	"dependencies": {
-		"@babel/runtime-corejs2": "7.12.5"
+		"@babel/runtime-corejs2": "7.12.5",
+		"@wordpress/components": "^11.1.1",
+		"@wordpress/compose": "^3.22.0",
+		"@wordpress/data": "^4.25.0",
+		"@wordpress/notices": "^2.11.0"
 	},
 	"publishConfig": {
 		"access": "public"
