+++ packages/customer-effort-score/src/index.js
@@ -3,16 +3,14 @@
  */
 import { useState } from '@wordpress/element';
 import PropTypes from 'prop-types';
-import {
-	Button,
-	Modal,
-	RadioControl,
-	__experimentalText as Text,
-	TextareaControl,
-} from '@wordpress/components';
 import { __ } from '@wordpress/i18n';
 import { compose } from '@wordpress/compose';
 import { withDispatch } from '@wordpress/data';
+
+/**
+ * Internal dependencies
+ */
+import Feedback from './feedback';
 
 /**
  * Use `CustomerEffortScore` to gather a customer effort score.
@@ -34,18 +32,8 @@ function CustomerEffortScore( {
 	openedCallback,
 	icon,
 } ) {
-	const [ score, setScore ] = useState();
-	const [ comments, setComments ] = useState();
 	const [ shouldCreateNotice, setShouldCreateNotice ] = useState( true );
 	const [ visible, setVisible ] = useState( false );
-
-	const closeModal = () => {
-		setVisible( false );
-	};
-	const sendScore = () => {
-		closeModal();
-		trackCallback( score, comments );
-	};
 
 	if ( shouldCreateNotice ) {
 		createNotice( 'success', label, {
@@ -73,70 +61,7 @@ function CustomerEffortScore( {
 		return null;
 	}
 
-	return (
-		<Modal
-			className="woocommerce-customer-effort-score"
-			title={ __( 'Please share your feedback', 'woocommerce-admin' ) }
-			onRequestClose={ closeModal }
-		>
-			<Text variant="subtitle.small" as="p">
-				{ label }
-			</Text>
-
-			<div className="woocommerce-customer-effort-score__selection">
-				<RadioControl
-					selected={ score }
-					options={ [
-						{
-							label: __( 'Very difficult', 'woocommerce-admin' ),
-							value: '1',
-						},
-						{
-							label: __(
-								'Somewhat difficult',
-								'woocommerce-admin'
-							),
-							value: '2',
-						},
-						{
-							label: __( 'Neutral', 'woocommerce-admin' ),
-							value: '3',
-						},
-						{
-							label: __( 'Somewhat easy', 'woocommerce-admin' ),
-							value: '4',
-						},
-						{
-							label: __( 'Very easy', 'woocommerce-admin' ),
-							value: '5',
-						},
-					] }
-					onChange={ ( value ) => setScore( value ) }
-				/>
-			</div>
-
-			{ ( score === '1' || score === '2' ) && (
-				<div className="woocommerce-customer-effort-score__comments">
-					<TextareaControl
-						label="Comments (Optional)"
-						help="Your feedback will go to the WooCommerce development team"
-						value={ comments }
-						onChange={ ( value ) => setComments( value ) }
-						rows="5"
-					/>
-				</div>
-			) }
-
-			<div className="woocommerce-customer-effort-score__buttons">
-				<Button isTertiary onClick={ closeModal }>
-					{ __( 'Cancel', 'woocommerce-admin' ) }
-				</Button>
-				<Button isPrimary onClick={ sendScore }>
-					{ __( 'Send', 'woocommerce-admin' ) }
-				</Button>
-			</div>
-		</Modal>
-	);
+	return <Feedback label={ label } trackCallback={ trackCallback } />;
 }
 
 CustomerEffortScore.propTypes = {
